modification de l'id des entity et effet de bord partourt (#88)

+++ src/main/java/com/joe/coffee/api/Dto/In/ConsommationDtoIn.java
@@ -12,8 +12,8 @@ public record ConsommationDtoIn(
         @NotNull
         Byte reglageIntensite,
         @NotNull
-        Long cafeId,
+        Integer cafeId,
         @NotNull
-        Long machineACafeId
+        Integer machineACafeId
 ) {
 }
