+++ src/main/java/com/joe/coffee/api/Exception/CafeExceptions/CafeCommercantNotFoundException.java
@@ -2,7 +2,7 @@ package com.joe.coffee.api.Exception.CafeExceptions;
 
 public class CafeCommercantNotFoundException extends RuntimeException {
 
-    public CafeCommercantNotFoundException(Long idComm) {
+    public CafeCommercantNotFoundException(Integer idComm) {
         super("Aucun café trouvé pour le commerçant avec id " + idComm);
     }
 }
