+++ src/main/java/com/joe/coffee/api/Exception/MarqueExceptions/MarqueNotFoundException.java
@@ -2,7 +2,7 @@ package com.joe.coffee.api.Exception.MarqueExceptions;
 
 public class MarqueNotFoundException extends RuntimeException {
 
-    public MarqueNotFoundException(Long id) {
+    public MarqueNotFoundException(Integer id) {
         super("Marque avec l'id " + id + " introuvable.");
     }
     public MarqueNotFoundException(String marque) {
